Refactoring and re-plumbing.
Things now... work (mostly)

Created generic 'panels', of which kiwi.model.Server and kiwi.model.Panel are sub-types of.

Server tab/panel is back, the MOTD and things not addressed to a channel that we're on go there.

Some issues remain. For some reason the topic is a bit dodgy. /part-ing doesn't work properly for instance.