Add IrcConnection.capEnd() function to avoid sending CAP END twice
Other than the correctness issue, this is needed to prevent a crash in
bitlbee 3.4.2 when CAP END is sent after registration, which kiwiirc
seems to trigger when it receives 900 followed by a 903 later.

The crash itself is fixed in git versions of bitlbee, but it would be
nice if kiwiirc worked fine with the last release.